refactor access request message processing #HB-28
