+++ src/main/java/com/jeksvp/bpd/service/access/AccessMessageProcessor.java
@@ -1,0 +1,8 @@
+package com.jeksvp.bpd.service.access;
+
+import com.jeksvp.bpd.kafka.dto.access.AccessRequestMsg;
+
+public interface AccessMessageProcessor {
+
+    void process(AccessRequestMsg accessRequestMsg);
+}
