+++ src/main/java/com/jeksvp/bpd/service/access/AccessService.java
@@ -1,4 +1,4 @@
-package com.jeksvp.bpd.service;
+package com.jeksvp.bpd.service.access;
 
 import com.jeksvp.bpd.kafka.dto.access.AccessRequestMsg;
 import com.jeksvp.bpd.web.dto.request.access.AccessRequest;
